+++ test/Skeletest/Internal/CLISpec.hs
@@ -1,47 +1,204 @@
 {-# LANGUAGE OverloadedRecordDot #-}
+{-# LANGUAGE OverloadedStrings #-}
 
 module Skeletest.Internal.CLISpec (spec) where
 
 import Control.Monad ((>=>))
 import Data.Dynamic (fromDynamic)
 import Data.Map qualified as Map
-import Data.Typeable (Typeable, typeOf)
+import Data.Typeable (typeOf)
 import Skeletest
-import Skeletest.Internal.CLI (
-  CLIFlagStore,
-  CLIParseResult (..),
-  flag,
-  parseCliArgs,
- )
+import Skeletest.Internal.CLI
 import Skeletest.Predicate qualified as P
 import Skeletest.TestUtils.Integration
 
-newtype FooFlag = FooFlag String
-  deriving (Eq)
-instance IsFlag FooFlag where
-  flagName = "foo"
-  flagHelp = "test"
-  flagSpec =
-    OptionalFlag
-      { flagDefault = FooFlag ""
-      , flagParse = Right . FooFlag
-      }
-
 spec :: Spec
 spec = do
-  describe "parseCliArgs" $ do
-    it "parses long flag" $ do
-      parseCliArgs [flag @FooFlag] ["--foo", "1"]
-        `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (FooFlag "1")}
+  spec_parseCliArgsWith
+  spec_getFlag
 
-    it "parses long flag with equal sign" $ do
-      parseCliArgs [flag @FooFlag] ["--foo=1"]
-        `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (FooFlag "1")}
+newtype MyFlag = MyFlag String
+  deriving (Eq)
+newtype MyFlag2 = MyFlag2 String
+  deriving (Eq)
 
-    it "parses long flag containing equal sign" $ do
-      parseCliArgs [flag @FooFlag] ["--foo=1=2"]
-        `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (FooFlag "1=2")}
+spec_parseCliArgsWith :: Spec
+spec_parseCliArgsWith = do
+  describe "parseCliArgsWith" $ do
+    longFlagSpec
+    shortFlagSpec
+    optFlagSpec
+    reqFlagSpec
+    switchFlagSpec
+    multiFlagSpec
+ where
+  mkFlagInfos name mShort fspec = [(name, mShort, SomeFlagSpec fspec)]
 
+  longFlagSpec = do
+    let flags =
+          mkFlagInfos "foo" Nothing $
+            RequiredFlag
+              { flagParse = pure . MyFlag
+              }
+    describe "long flag" $ do
+      it "parses" $ do
+        parseCliArgsWith flags ["--foo", "1"]
+          `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (MyFlag "1")}
+      it "parses with equal sign" $ do
+        parseCliArgsWith flags ["--foo=1"]
+          `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (MyFlag "1")}
+      it "parses argument containing equal sign" $ do
+        parseCliArgsWith flags ["--foo=1=2"]
+          `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (MyFlag "1=2")}
+      it "errors if unknown" $ do
+        parseCliArgsWith [] ["--foo"]
+          `shouldSatisfy` parseFailure "Unknown flag: --foo"
+
+  shortFlagSpec = do
+    let flags =
+          mkFlagInfos "foo" (Just 'f') $
+            RequiredFlag
+              { flagParse = pure . MyFlag
+              }
+    describe "short flag" $ do
+      it "parses" $ do
+        parseCliArgsWith flags ["-f", "123"]
+          `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (MyFlag "123")}
+      it "parses multiple switches at once" $ do
+        let flags' =
+              [ ("flag-a", Just 'a', SomeFlagSpec $ SwitchFlag (MyFlag . show))
+              , ("flag-b", Just 'b', SomeFlagSpec $ SwitchFlag (MyFlag2 . show))
+              ]
+            expected =
+              P.and
+                [ containsFlag (MyFlag "True")
+                , containsFlag (MyFlag2 "True")
+                ]
+        parseCliArgsWith flags' ["-ab"]
+          `shouldSatisfy` P.con CLIParseSuccess{flagStore = expected}
+      it "parses short flag with arg without space" $ do
+        let flags' =
+              mkFlagInfos "foo" (Just 'f') $
+                RequiredFlag
+                  { flagParse = pure . MyFlag
+                  }
+        parseCliArgsWith flags' ["-fasdf"]
+          `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (MyFlag "asdf")}
+      it "errors if unknown" $ do
+        parseCliArgsWith [] ["-x"]
+          `shouldSatisfy` parseFailure "Unknown flag: -x"
+
+  optFlagSpec = do
+    let flags =
+          mkFlagInfos "foo" Nothing $
+            OptionalFlag
+              { flagDefault = MyFlag ""
+              , flagParse = pure . MyFlag
+              }
+    describe "OptionalFlag" $ do
+      it "returns last flag" $ do
+        parseCliArgsWith flags ["--foo", "1", "--foo", "2"]
+          `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (MyFlag "2")}
+      it "returns default if not set" $ do
+        parseCliArgsWith flags []
+          `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (MyFlag "")}
+      it "errors if no argument" $ do
+        parseCliArgsWith flags ["--foo"]
+          `shouldSatisfy` parseFailure "Flag '--foo' requires argument"
+
+  reqFlagSpec = do
+    let flags =
+          mkFlagInfos "foo" Nothing $
+            RequiredFlag
+              { flagParse = pure . MyFlag
+              }
+    describe "RequiredFlag" $ do
+      it "returns last flag" $ do
+        parseCliArgsWith flags ["--foo", "1", "--foo", "2"]
+          `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (MyFlag "2")}
+      it "errors if not set" $ do
+        parseCliArgsWith flags []
+          `shouldSatisfy` parseFailure "Flag '--foo' is required"
+      it "errors if no argument" $ do
+        parseCliArgsWith flags ["--foo"]
+          `shouldSatisfy` parseFailure "Flag '--foo' requires argument"
+
+  switchFlagSpec = do
+    let flags =
+          mkFlagInfos "foo" Nothing $
+            SwitchFlag
+              { flagFromBool = MyFlag . show
+              }
+    describe "SwitchFlag" $ do
+      it "returns True if set" $ do
+        parseCliArgsWith flags ["--foo"]
+          `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (MyFlag "True")}
+      it "returns True if set any number of times" $ do
+        parseCliArgsWith flags ["--foo", "--foo", "--foo"]
+          `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (MyFlag "True")}
+      it "returns False if not set" $ do
+        parseCliArgsWith flags []
+          `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (MyFlag "False")}
+      it "errors if argument is set" $ do
+        parseCliArgsWith flags ["--foo=asdf"]
+          `shouldSatisfy` parseFailure "Flag '--foo' does not take arguments, got: asdf"
+
+  multiFlagSpec = do
+    let mkMultiFlags :: (Show a) => MultiFlagType a -> FlagInfos
+        mkMultiFlags type_ =
+          mkFlagInfos "foo" (Just 'f') $
+            MultiFlag
+              { type_
+              , parseMulti = pure . MyFlag . show
+              }
+    describe "MultiFlag" $ do
+      describe "ManyFlag" $ do
+        let flags = mkMultiFlags $ ManyFlag FlagType_Arg
+        it "parses none" $ do
+          parseCliArgsWith flags []
+            `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (MyFlag "[]")}
+        it "parses one" $ do
+          parseCliArgsWith flags ["--foo", "1"]
+            `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (MyFlag "[\"1\"]")}
+        it "parses multiple" $ do
+          parseCliArgsWith flags ["--foo", "1", "--foo", "2"]
+            `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (MyFlag "[\"1\",\"2\"]")}
+      describe "SomeFlag" $ do
+        let flags = mkMultiFlags $ SomeFlag FlagType_Arg
+        it "errors if not provided" $ do
+          parseCliArgsWith flags []
+            `shouldSatisfy` parseFailure "Flag '--foo' is required"
+        it "parses one" $ do
+          parseCliArgsWith flags ["--foo", "1"]
+            `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (MyFlag "\"1\" :| []")}
+        it "parses multiple" $ do
+          parseCliArgsWith flags ["--foo", "1", "--foo", "2"]
+            `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (MyFlag "\"1\" :| [\"2\"]")}
+      describe "FlagType_Switch" $ do
+        let flags = mkMultiFlags $ ManyFlag FlagType_Switch
+        it "parses" $ do
+          parseCliArgsWith flags ["--foo", "--foo"]
+            `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (MyFlag "[True,True]")}
+        it "parses multiple short flags" $ do
+          parseCliArgsWith flags ["-fff"]
+            `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (MyFlag "[True,True,True]")}
+        it "errors if argument provided" $ do
+          parseCliArgsWith flags ["--foo=asdf"]
+            `shouldSatisfy` parseFailure "Flag '--foo' does not take arguments, got: asdf"
+      describe "FlagType_Arg" $ do
+        let flags = mkMultiFlags $ ManyFlag FlagType_Arg
+        it "parses" $ do
+          parseCliArgsWith flags ["--foo", "1", "--foo", "2"]
+            `shouldSatisfy` P.con CLIParseSuccess{flagStore = containsFlag (MyFlag "[\"1\",\"2\"]")}
+        it "errors if no argument" $ do
+          parseCliArgsWith flags ["--foo"]
+            `shouldSatisfy` parseFailure "Flag '--foo' requires argument"
+
+  containsFlag f = (Map.lookup (typeOf f) >=> fromDynamic) P.>>> P.just (P.eq f)
+  parseFailure msg = P.con (CLIParseFailure (P.eq msg))
+
+spec_getFlag :: Spec
+spec_getFlag = do
   describe "getFlag" $ do
     integration . it "reads registered flag" $ do
       runner <- getFixture @TestRunner
@@ -89,6 +246,3 @@ spec = do
       (stdout, stderr) <- expectFailure runner.runTests
       stderr `shouldBe` ""
       stdout `shouldSatisfy` P.matchesSnapshot
-
-containsFlag :: (Typeable a, Eq a) => a -> Predicate IO CLIFlagStore
-containsFlag f = (Map.lookup (typeOf f) >=> fromDynamic) P.>>> P.just (P.eq f)